refactor: coyote.http11 패키지 정리
