+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -2,12 +2,12 @@ package org.apache.coyote.http11;
 
 import camp.nextstep.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http11.request.RequestParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.*;
 import java.net.Socket;
-import java.nio.charset.StandardCharsets;
 
 public class Http11Processor implements Runnable, Processor {
