+++ tomcat/src/main/java/org/apache/coyote/http11/NotFoundHandler.java
@@ -1,6 +1,11 @@
 package org.apache.coyote.http11;
 
 
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.response.HttpResponseHeaders;
+import org.apache.coyote.http11.response.HttpStatusCode;
+
 public class NotFoundHandler implements RequestHandler {
 
     public static final RequestHandler INSTANCE = new NotFoundHandler();
