+++ tomcat/src/main/java/org/apache/coyote/http11/StaticResourceRequestHandler.java
@@ -1,6 +1,11 @@
 package org.apache.coyote.http11;
 
 import camp.nextstep.exception.UncheckedServletException;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.response.HttpResponseHeaders;
+import org.apache.coyote.http11.response.HttpStatusCode;
+import org.apache.coyote.http11.response.ResponseBody;
 
 import java.io.IOException;
