+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequestLineInvalidException.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 public class HttpRequestLineInvalidException extends RuntimeException {
     public HttpRequestLineInvalidException(String msg) {
