+++ tomcat/src/main/java/org/apache/coyote/http11/request/RequestLine.java
@@ -1,5 +1,7 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
+
+import org.apache.coyote.http11.HttpProtocol;
 
 import java.util.Map;
