+++ tomcat/src/main/java/org/apache/coyote/http11/request/RequestParser.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 
 import camp.nextstep.exception.UncheckedServletException;
@@ -8,7 +8,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.nio.charset.Charset;
-import java.sql.Array;
 import java.util.ArrayList;
 import java.util.List;
