+++ tomcat/src/main/java/org/apache/coyote/http11/response/Location.java
@@ -1,4 +1,6 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.response;
+
+import org.apache.coyote.http11.HttpProtocol;
 
 public record Location(String url) {
