+++ tomcat/src/main/java/org/apache/coyote/http11/response/ResponseBody.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.response;
 
 import java.nio.charset.Charset;
